+++ CHANGES.md
@@ -6,6 +6,7 @@ Changes
 
 ### Unreleased
 
+* 2018-07-31 - Fixed bug in function theme_boost_campus_process_flatnav.
 * 2018-07-31 - Removed deprecated.txt file because it is unnecessary.
 * 2018-07-24 - Setting to display help texts in modal dialogues instead of popups.
 * 2018-07-24 - Fixed brand colors due to Bootstrap changes in Boost.